requirements.txt
streamlit
torch
torchvision
opencv-python-headless
pandas
openpyxl
numpy
Pillow
